# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Flappy Bird Backend API
  version: 1.0.0
  summary: Optional accounts, leaderboards, achievements and score sync for the Flappy Bird iOS game.
  description: |
    # Overview

    This service is the **optional** companion backend for the
    [Flappy Bird iOS game](https://github.com/hoangsonww/Flappy-Bird-Game).

    The game is fully playable without it. When a server is reachable, the app
    unlocks accounts, cloud-synced progress, global/daily/weekly leaderboards,
    friend boards, achievements and the deterministic daily challenge.

    ## Discovery handshake

    On launch the game probes a short list of candidate base URLs and calls
    `GET /v1/meta/config`. A response containing `"protocol": "flappy-bird/1"`
    switches the app into online mode; anything else keeps it offline. Nothing
    else about the app's behaviour depends on the server being present.

    ```mermaid
    sequenceDiagram
      participant Game as iOS Game
      participant API as Backend
      Game->>API: GET /v1/meta/config
      API-->>Game: 200 { protocol, capabilities, limits }
      Game->>API: POST /v1/auth/guest  { deviceId }
      API-->>Game: 201 { user, tokens }
      Game->>API: POST /v1/scores      (Bearer access token)
      API-->>Game: 201 { score, rank, unlockedAchievements }
    ```

    ## Authentication

    Two token types:

    | Token         | Lifetime            | Storage on device | Purpose                        |
...

...

...
    history.

...

...

...
    {
...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/scores/{id}:
    get:
      tags: [Scores]
      summary: Fetch one run
      description: Owners and admins only.
      operationId: getScoreById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The run
          content:
            application/json:
              schema:
                type: object
                properties:
                  score: { $ref: '#/components/schemas/Score' }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/admin/stats:
    get:
      tags: [Admin]
      summary: Instance counters
      security:
        - bearerAuth: []
        - adminToken: []
      operationId: getAdminStats
      responses:
        '200':
          description: Counters
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: { type: integer }
                  runs: { type: integer }
                  totals:
                    type: object
                    properties:
                      players: { type: integer }
                      games: { type: integer }
                      pipes: { type: integer }
                      coins: { type: integer }
                  driver: { $ref: '#/components/schemas/StorageDriver' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/admin/users/{username}/ban:
    post:
      tags: [Admin]
      summary: Ban a player
      description: Revokes every session and removes the player from all boards.
      security:
        - bearerAuth: []
        - adminToken: []
      operationId: banUser
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  { type: string, minLength: 3, maxLength: 200, default: Violated fair play rules }
      responses:
        '200':
          description: Player banned
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      username: { type: string }
                      isBanned: { type: boolean }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin]
      summary: Lift a ban
      security:
        - bearerAuth: []
        - adminToken: []
      operationId: unbanUser
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
      responses:
        '200':
          description: Ban lifted
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      username: { type: string }
                      isBanned: { type: boolean }
  /v1/admin/scores/{id}/flag:
    post:
      tags: [Admin]
      summary: Flag a run
      description: Hides the run and recomputes the owner's aggregate stats.
      security:
        - bearerAuth: []
        - adminToken: []
      operationId: flagScore
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 3, maxLength: 200 }
      responses:
        '200':
          description: Run flagged
          content:
            application/json:
              schema:
                type: object
                properties:
                  flagged: { type: boolean }
                  id: { type: string, format: uuid }
                  reason: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/admin/scores/{id}:
    delete:
      tags: [Admin]
      summary: Delete a run
      security:
        - bearerAuth: []
        - adminToken: []
      operationId: deleteScore
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Run deleted }
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/admin/maintenance/purge-tokens:
    post:
      tags: [Admin]
      summary: Purge expired refresh tokens
      security:
        - bearerAuth: []
        - adminToken: []
      operationId: purgeTokens
      responses:
        '200':
          description: Number of rows removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  purged: { type: integer }
  /v1/admin/maintenance/recompute-stats/{username}:
    post:
      tags: [Admin]
      summary: Recompute a player's aggregate stats
      description: Rebuilds `user_stats` from the raw run table. Useful after moderation.
      security:
        - bearerAuth: []
        - adminToken: []
      operationId: recomputeStats
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
      responses:
        '200':
          description: Recomputed stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats: { $ref: '#/components/schemas/UserStats' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token from `/v1/auth/login`, `/register`, `/guest` or `/refresh`.
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Static token matching the server's `ADMIN_TOKEN`. An alternative to an admin account.
  parameters:
    UsernamePath:
      name: username
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Username' }
...
  schemas:
    Username:
      type: string
      pattern: '^[a-zA-Z0-9_]{3,20}$'
      example: skyhopper
      description: 3–20 characters; letters, digits and underscore. Compared case-insensitively.
...
    GameMode:
      type: string
      enum: [classic, endless, timeAttack, hardcore, zen, daily]
      description: |
        * `classic` — the original rules.
        * `endless` — difficulty ramps continuously.
        * `timeAttack` — score as much as possible in 60 seconds.
        * `hardcore` — narrow gaps, one life, no power-ups.
        * `zen` — no death, for practice (excluded from most boards by convention).
        * `daily` — today's deterministic challenge.
...
    StorageDriver:
      type: string
      enum: [postgres, memory]
...
    Score:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        score: { type: integer, minimum: 0 }
        mode: { $ref: '#/components/schemas/GameMode' }
        coins: { type: integer, minimum: 0 }
        pipesPassed: { type: integer, minimum: 0 }
        durationMs: { type: integer, minimum: 0 }
        maxCombo: { type: integer, minimum: 0 }
        powerUpsUsed: { type: integer, minimum: 0 }
        seed: { type: string, description: Run seed; lets the game replay the same pipe layout. }
        clientVersion: { type: [string, 'null'] }
        deviceModel: { type: [string, 'null'] }
        flagged: { type: boolean, description: Flagged runs are excluded from boards and stats. }
        flagReason: { type: [string, 'null'] }
        submittedAt: { type: string, format: date-time }
...
    UserStats:
      type: object
      properties:
        userId: { type: string, format: uuid }
        bestScore: { type: integer }
        bestScoreMode: { oneOf: [{ $ref: '#/components/schemas/GameMode' }, { type: 'null' }] }
        totalScore: { type: integer }
        gamesPlayed: { type: integer }
        totalCoins: { type: integer }
        totalPipes: { type: integer }
        totalDurationMs: { type: integer }
        longestRunMs: { type: integer }
        bestCombo: { type: integer }
        updatedAt: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - validation_failed
                - unauthorized
                - invalid_credentials
                - token_expired
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - payload_too_large
                - unprocessable
                - service_unavailable
                - internal_error
            message: { type: string }
            details: { description: Structured context; shape depends on `code`. }
            requestId: { type: string, description: Matches the `X-Request-Id` response header. }
  responses:
...
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...